Graphics: Improve billboard system
The billboard system singleton was removed. The billboard system
object is now accessible via the GFX singleton.
