+++ game/src/game/Graphics/RenderPasses.cpp
@@ -26,7 +26,7 @@
 #include "game/graphic_prt.h"
 #include "game/graphic_mad.h"
 #include "game/graphic_fan.h"
-#include "game/graphic_billboard.h"
+#include "game/Graphics/BillboardSystem.hpp"
 #include "game/renderer_3d.h"
 #include "game/Logic/Player.hpp"
 #include "egolib/Script/script.h"
